// types/Pokemon.types.ts
export interface Pokemon {
  [x: string]: any
  id: number
  height: number
  name: string
  sprites: Sprites
  stats: Stat[]
  weight: number
}
// ...
export interface Sprites {
  front_default: string
}
// ...
export interface Stat {
  base_stat: number
  effort: number
  stat: Stat2
}

export interface Stat2 {
  name: string
  url: string
}